moving mobs working

+++ game_server/src/long_term_storage_service/db_character.rs
@@ -9,7 +9,7 @@ pub struct StoredCharacter {
     #[serde(rename = "_id", skip_serializing_if = "Option::is_none")]
     pub id: Option<ObjectId>,
     pub world_id: Option<ObjectId>,
-    pub player_id: u64,
+    pub player_id: u16,
     pub faction: String,
     pub device_id: String,
     pub character_name: String,
