+++ game_server/src/long_term_storage_service/db_player.rs
@@ -6,7 +6,7 @@ use serde::{Serialize, Deserialize};
 pub struct StoredPlayer {
     #[serde(rename = "_id", skip_serializing_if = "Option::is_none")]
     pub id: Option<ObjectId>,
-    pub player_id: u64,
+    pub player_id: u16,
     pub device_id:String,
     pub player_name: String,
     pub constitution: u32
